Update arithmetic_arranger.py
The original code does not recognize a situation where the problems contain only '+' or only '-' as the operator. Although this condition is completely fair, the code gives an error message. Now, it will recognize length = 1 with either + or - as a legitimate scenario and will continue the function rather than returning error.